# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Streaming de Cursos",
        "description": "API para gerenciamento de cursos, usuários e assinaturas. Para usar a API, faça autenticação pelo endpoint /api/login ou usando OAuth2.",
        "contact": {
            "name": "Suporte API",
            "email": "contato@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/subscriptions/update-payment-method": {
            "post": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Atualizar método de pagamento",
                "description": "Atualiza o método de pagamento de uma assinatura",
                "operationId": "updatePaymentMethod",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subscription_id",
                                    "payment_method"
                                ],
                                "properties": {
                                    "subscription_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "pm_card_visa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Método de pagamento atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Método de pagamento atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido - assinatura não pertence ao usuário"
                    },
                    "404": {
                        "description": "Assinatura não encontrada"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/subscriptions/{subscription}/cancel-stripe": {
            "post": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Cancelar assinatura no Stripe",
                "description": "Cancela uma assinatura no sistema e no Stripe",
                "operationId": "cancelStripeSubscription",
                "parameters": [
                    {
                        "name": "subscription",
                        "in": "path",
                        "description": "ID da assinatura",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assinatura cancelada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Assinatura cancelada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido - assinatura não pertence ao usuário"
                    },
                    "404": {
                        "description": "Assinatura não encontrada"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/subscriptions/setup-intent": {
            "get": {
                "tags": [
                    "Assinaturas"
                ],
                "summary": "Criar intent de configuração para cartão",
                "description": "Cria um intent de configuração do Stripe para salvar dados de cartão",
                "operationId": "createSetupIntent",
                "responses": {
                    "200": {
                        "description": "Intent de configuração criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "client_secret": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "oauth2": {
                "type": "oauth2",
                "description": "OAuth2 para autenticação de API",
                "flows": {
                    "password": {
                        "tokenUrl": "http://localhost/oauth/token",
                        "scopes": []
                    }
                }
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}